+++ sqlsense/parser.py
@@ -51,9 +51,9 @@ class SqlParser(object):
                 rule_key = 'Token.Literal.String'
             else:
                 rule_key = str(tk.ttype)
-            action = self._parse_rules[rule_key] if rule_key in self._parse_rules else None
+            (action, tokengrp_set) = self._parse_rules[rule_key] if rule_key in self._parse_rules else (None, None)
             if action:
-                curr_tk_grp = action(stmt, curr_tk_grp, tk)
+                curr_tk_grp = action(stmt, curr_tk_grp, tk, tokengrp_set)
             else:
                 curr_tk_grp.append(tk)
 
@@ -86,11 +86,25 @@ class SqlParser(object):
         token_group.append(token)
         return token_group
 
-    def _process_select_clause(self, stmt, token_group, token):
+    def _process_select_clause(self, stmt, token_group, token, tokengroup_set):
         select_clause_grp = TokenGroup([token, ], ST.SelectClause)
-        if token_group.ttype is not None and token_group.ttype not in (ST.InsertIntoClause, ST.RoundBracket, ST.CollectionSet):
-            # If Select Clause is preceeded by Insert Into Clause, get out of that Clause Token Group
-            token_group = self._switch_to_parent(token_group)
+        if token_group.ttype in (ST.RoundBracket, ST.CollectionSet):
+            # TODO: Should we check the preceeding token to make sure SELECT is the first keyword in Brackets
+            token_group.ttype = ST.SubQuery
+        elif (token_group.ttype in (ST.Select, ST.SelectInto, ST.InsertIntoSelect, ST.SubQuery)
+              and (token_group.last_token().match_type_value(Token(T.Keyword, 'UNION'))
+                   or (token_group.last_token().match_type_value(Token(T.Keyword, 'ALL'))
+                       and token_group.token_before(next_token_index=token_group.last_token_index()).match_type_value(Token(T.Keyword, 'UNION'))
+                       )
+                   )
+              ):  # token_group.has_token_as_immediate_child(Token(T.Keyword, 'UNION')):
+            pass
+        else:
+            while token_group.ttype is not None and token_group.ttype not in tokengroup_set:
+                # If Select Clause is preceeded by Insert Into Clause, get out of that Clause Token Group
+                token_group = self._switch_to_parent(token_group)
+            if token_group.ttype in tokengroup_set:
+                token_group = self._switch_to_parent(token_group)
 
         token_group.append(select_clause_grp)
 
@@ -100,12 +114,9 @@ class SqlParser(object):
                 token_group.ttype = ST.Select
             elif token_group.ttype is ST.Insert:
                 token_group.ttype = ST.InsertIntoSelect
-        elif token_group.ttype in (ST.RoundBracket, ST.CollectionSet):
-            # TODO: Should we check the preceeding token to make sure SELECT is the first keyword in Brackets
-            token_group.ttype = ST.SubQuery
         return select_clause_grp
 
-    def _process_from_clause(self, stmt, token_group, token):
+    def _process_from_clause(self, stmt, token_group, token, tokengroup_set):
         from_clause_grp = TokenGroup([token, ], ST.FromClause)
 
         while token_group.ttype not in (ST.SelectClause, ST.SelectIntoClause, ST.UpdateSetClause):
@@ -120,7 +131,7 @@ class SqlParser(object):
         # TODO: Think about Delete From Clause before returning
         return from_clause_grp
 
-    def _process_where_clause(self, stmt, token_group, token):
+    def _process_where_clause(self, stmt, token_group, token, tokengroup_set):
         where_clause_grp = TokenGroup([token, ], ST.WhereClause)
 
         while token_group.ttype not in (ST.FromClause, ):
@@ -132,7 +143,7 @@ class SqlParser(object):
         token_group.append(where_clause_grp)
         return where_clause_grp
 
-    def _process_group_by_clause(self, stmt, token_group, token):
+    def _process_group_by_clause(self, stmt, token_group, token, tokengroup_set):
         group_by_clause_grp = TokenGroup([token, ], ST.GroupByClause)
 
         while token_group.ttype not in (ST.FromClause, ST.WhereClause):
@@ -144,7 +155,7 @@ class SqlParser(object):
         token_group.append(group_by_clause_grp)
         return group_by_clause_grp
 
-    def _process_having_clause(self, stmt, token_group, token):
+    def _process_having_clause(self, stmt, token_group, token, tokengroup_set):
         having_clause_grp = TokenGroup([token, ], ST.HavingClause)
 
         while token_group.ttype not in (ST.FromClause, ST.WhereClause, ST.GroupByClause):
@@ -157,7 +168,7 @@ class SqlParser(object):
         token_group.append(having_clause_grp)
         return having_clause_grp
 
-    def _process_order_by_clause(self, stmt, token_group, token):
+    def _process_order_by_clause(self, stmt, token_group, token, tokengroup_set):
         order_by_clause_grp = TokenGroup([token, ], ST.OrderByClause)
 
         while token_group.ttype not in (ST.FromClause, ST.WhereClause, ST.GroupByClause, ST.HavingClause):
@@ -169,14 +180,14 @@ class SqlParser(object):
         token_group.append(order_by_clause_grp)
         return order_by_clause_grp
 
-    def _process_join(self, stmt, token_group, token):
+    def _process_join(self, stmt, token_group, token, tokengroup_set):
         while token_group.ttype != ST.FromClause:
             # Get out of the Token Group until you find the matching From Clause
             token_group = self._switch_to_parent(token_group)
         token_group.append(token)
         return token_group
 
-    def _process_on(self, stmt, token_group, token):
+    def _process_on(self, stmt, token_group, token, tokengroup_set):
         join_on_grp = TokenGroup([token, ], ST.JoinOnClause)
         while token_group.ttype != ST.FromClause:
             # Get out of the Token Group until you find the matching From Clause
@@ -184,7 +195,7 @@ class SqlParser(object):
         token_group.append(join_on_grp)
         return join_on_grp
 
-    def _process_and(self, stmt, token_group, token):
+    def _process_and(self, stmt, token_group, token, tokengroup_set):
         token.ttype = ST.LogicalOperator
         while token_group.ttype not in (ST.JoinOnClause, ST.WhereClause, ST.HavingClause,
                                         ST.ConditionGroup, ST.Between, ST.NotBetween,
@@ -195,11 +206,11 @@ class SqlParser(object):
             # Between Clause should not have more than one AND Operator
             # Get out of the Token Group until you find preceeding Where, Having Clause
             token_group = self._switch_to_parent(token_group)
-            return self._process_and(stmt, token_group, token)
+            return self._process_and(stmt, token_group, token, tokengroup_set)
         token_group.append(token)
         return token_group
 
-    def _process_or(self, stmt, token_group, token):
+    def _process_or(self, stmt, token_group, token, tokengroup_set):
         token.ttype = ST.LogicalOperator
         while token_group.ttype not in (ST.JoinOnClause, ST.WhereClause, ST.HavingClause,
                                         ST.ConditionGroup, ST.Between, ST.NotBetween,
@@ -209,7 +220,7 @@ class SqlParser(object):
         token_group.append(token)
         return token_group
 
-    def _process_in(self, stmt, token_group, token):
+    def _process_in(self, stmt, token_group, token, tokengroup_set):
         while token_group.ttype not in (ST.JoinOnClause, ST.WhereClause, ST.HavingClause,
                                         ST.Condition, ST.RoundBracket, ST.ConditionGroup, ST.Not,
                                         ST.CaseExpression, ST.WhenExpression, ST.ThenExpression, ST.ElseExpression):
@@ -227,25 +238,25 @@ class SqlParser(object):
         token_group.append(token)
         return token_group
 
-    def _process_like(self, stmt, token_group, token):
+    def _process_like(self, stmt, token_group, token, tokengroup_set):
         # LIKE is same as IN, just that the ttype should be Like/NotLike
-        token_group = self._process_in(stmt, token_group, token)
+        token_group = self._process_in(stmt, token_group, token, tokengroup_set)
         token_group.ttype = ST.Like if token_group.ttype == ST.In else ST.NotLike
         return token_group
 
-    def _process_between(self, stmt, token_group, token):
+    def _process_between(self, stmt, token_group, token, tokengroup_set):
         # BETWEEN is same as IN, just that the ttype should be Between/NotBetween
-        token_group = self._process_in(stmt, token_group, token)
+        token_group = self._process_in(stmt, token_group, token, tokengroup_set)
         token_group.ttype = ST.Between if token_group.ttype == ST.In else ST.NotBetween
         return token_group
 
-    def _process_is(self, stmt, token_group, token):
+    def _process_is(self, stmt, token_group, token, tokengroup_set):
         # IS is same as IN, just that the ttype should be Comparison
-        token_group = self._process_in(stmt, token_group, token)
+        token_group = self._process_in(stmt, token_group, token, tokengroup_set)
         token_group.ttype = ST.Comparison
         return token_group
 
-    def _process_into(self, stmt, token_group, token):
+    def _process_into(self, stmt, token_group, token, tokengroup_set):
         # Assumption: INTO Clause will never be in a Sub-Query
         while token_group.ttype not in (ST.InsertIntoClause, ST.SelectClause):
             # Get out of the Token Group until you find preceeding
@@ -264,7 +275,7 @@ class SqlParser(object):
 
         return token_group
 
-    def _process_not(self, stmt, token_group, token):
+    def _process_not(self, stmt, token_group, token, tokengroup_set):
         # TODO: NOT can be in SELECT Clause as well
         # TODO: IS Condition
         token.ttype = ST.LogicalOperator
@@ -286,18 +297,18 @@ class SqlParser(object):
             token_group.append(token)
             return token_group
 
-    def _process_as(self, stmt, token_group, token):
+    def _process_as(self, stmt, token_group, token, tokengroup_set):
         if token_group.ttype == ST.SelectClause:
             # Case: SELECT CASE ... END AS case_alias
             token_group = token_group.merge_into_token_group(
                 ST.ComputedIdentifier, token_list_start_index_included=token_group.last_token_index())
         else:
-            while token_group.parent.ttype not in (ST.SelectClause, ST.FromClause):
+            while token_group.parent.ttype not in tokengroup_set:
                 token_group = self._switch_to_parent(token_group)
         token_group.append(token)
         return token_group
 
-    def _process_case(self, stmt, token_group, token):
+    def _process_case(self, stmt, token_group, token, tokengroup_set):
         case_exp_grp = TokenGroup([token, ], ST.CaseExpression)
         while token_group.ttype not in (ST.SelectClause, ST.JoinOnClause, ST.WhereClause, ST.HavingClause,
                                         ST.ConditionGroup, ST.Condition, ST.RoundBracket, ST.Not,
@@ -308,7 +319,7 @@ class SqlParser(object):
         token_group.append(case_exp_grp)
         return case_exp_grp
 
-    def _process_when(self, stmt, token_group, token):
+    def _process_when(self, stmt, token_group, token, tokengroup_set):
         when_exp_grp = TokenGroup([token, ], ST.WhenExpression)
         while token_group.ttype not in (ST.CaseExpression):
             # Get out of the Token Group until you find Case Expression
@@ -316,7 +327,7 @@ class SqlParser(object):
         token_group.append(when_exp_grp)
         return when_exp_grp
 
-    def _process_then(self, stmt, token_group, token):
+    def _process_then(self, stmt, token_group, token, tokengroup_set):
         then_exp_grp = TokenGroup([token, ], ST.ThenExpression)
         while token_group.ttype not in (ST.WhenExpression):
             # Get out of the Token Group until you find When Expression
@@ -324,7 +335,7 @@ class SqlParser(object):
         token_group.append(then_exp_grp)
         return then_exp_grp
 
-    def _process_else(self, stmt, token_group, token):
+    def _process_else(self, stmt, token_group, token, tokengroup_set):
         else_exp_grp = TokenGroup([token, ], ST.ElseExpression)
         while token_group.ttype not in (ST.CaseExpression):
             # Get out of the Token Group until you find Case Expression
@@ -332,7 +343,7 @@ class SqlParser(object):
         token_group.append(else_exp_grp)
         return else_exp_grp
 
-    def _process_end(self, stmt, token_group, token):
+    def _process_end(self, stmt, token_group, token, tokengroup_set):
         while token_group.ttype not in (ST.CaseExpression):
             # Get out of the Token Group until you find Case Expression
             token_group = self._switch_to_parent(token_group)
@@ -340,7 +351,14 @@ class SqlParser(object):
         token_group = self._switch_to_parent(token_group)
         return token_group
 
-    def _process_literal_number(self, stmt, token_group, token):
+    def _process_union(self, stmt, token_group, token, tokengroup_set):
+        while token_group.ttype not in (ST.Select, ST.SelectInto, ST.InsertIntoSelect, ST.SubQuery):
+            # Get out of the Token Group until you find Case Expression
+            token_group = self._switch_to_parent(token_group)
+        token_group.append(token)
+        return token_group
+
+    def _process_literal_number(self, stmt, token_group, token, tokengroup_set):
         if token_group.ttype == ST.SelectClause:
             select_constant_identifier_grp = TokenGroup(
                 [token, ], ST.SelectConstantIdentifier)
@@ -350,7 +368,7 @@ class SqlParser(object):
             token_group.append(token)
         return token_group
 
-    def _process_literal_string(self, stmt, token_group, token):
+    def _process_literal_string(self, stmt, token_group, token, tokengroup_set):
         token.ttype = T.String
         if token_group.ttype == ST.SelectClause:
             select_constant_identifier_grp = TokenGroup(
@@ -364,55 +382,56 @@ class SqlParser(object):
             token_group.append(token)
         return token_group
 
-    def _process_name(self, stmt, token_group, token):
+    def _process_name(self, stmt, token_group, token, tokengroup_set):
         raise NotImplementedError
 
-    def _process_punctuation(self, stmt, token_group, token):
+    def _process_punctuation(self, stmt, token_group, token, tokengroup_set):
         raise NotImplementedError
 
-    def _process_operator(self, stmt, token_group, token):
+    def _process_operator(self, stmt, token_group, token, tokengroup_set):
         raise NotImplementedError
 
     def _set_rules_(self):
         ''' You may extend this function in Child classes
         '''
         return {
-            'Token.Comment.Single': None,
-            'Token.Comment.Multiline': None,
-            'Token.Text.Whitespace': None,
-            'Token.Text.Name': self._process_name,
-            'Token.Text.Punctuation': self._process_punctuation,
-            'Token.Text.Operator': self._process_operator,
-            'Token.Keyword.SELECT': self._process_select_clause,
-            'Token.Keyword.DISTINCT': None,
-            'Token.Keyword.FROM': self._process_from_clause,
-            'Token.Keyword.WHERE': self._process_where_clause,
-            'Token.Keyword.GROUP': self._process_group_by_clause,
-            'Token.Keyword.HAVING': self._process_having_clause,
-            'Token.Keyword.ORDER': self._process_order_by_clause,
-            'Token.Keyword.BY': None,
-            'Token.Keyword.JOIN': self._process_join,
-            'Token.Keyword.INNER': self._process_join,
-            'Token.Keyword.LEFT': self._process_join,
-            'Token.Keyword.RIGHT': self._process_join,
-            'Token.Keyword.FULL': self._process_join,
-            'Token.Keyword.OUTER': self._process_join,
-            'Token.Keyword.CROSS': self._process_join,
-            'Token.Keyword.ON': self._process_on,
-            'Token.Keyword.AND': self._process_and,
-            'Token.Keyword.OR': self._process_or,
-            'Token.Keyword.IN': self._process_in,
-            'Token.Keyword.LIKE': self._process_like,
-            'Token.Keyword.BETWEEN': self._process_between,
-            'Token.Keyword.IS': self._process_is,
-            'Token.Keyword.AS': self._process_as,
-            'Token.Keyword.INTO': self._process_into,
-            'Token.Keyword.NOT': self._process_not,
-            'Token.Keyword.CASE': self._process_case,
-            'Token.Keyword.WHEN': self._process_when,
-            'Token.Keyword.THEN': self._process_then,
-            'Token.Keyword.ELSE': self._process_else,
-            'Token.Keyword.END': self._process_end,
-            'Token.Literal.Number': self._process_literal_number,
-            'Token.Literal.String': self._process_literal_string,
+            'Token.Comment.Single': (None, None),
+            'Token.Comment.Multiline': (None, None),
+            'Token.Text.Whitespace': (None, None),
+            'Token.Text.Name': (self._process_name, None),
+            'Token.Text.Punctuation': (self._process_punctuation, None),
+            'Token.Text.Operator': (self._process_operator, None),
+            'Token.Keyword.SELECT': (self._process_select_clause, (ST.InsertIntoClause, )),
+            'Token.Keyword.DISTINCT': (None, None),
+            'Token.Keyword.FROM': (self._process_from_clause, None),
+            'Token.Keyword.WHERE': (self._process_where_clause, None),
+            'Token.Keyword.GROUP': (self._process_group_by_clause, None),
+            'Token.Keyword.HAVING': (self._process_having_clause, None),
+            'Token.Keyword.ORDER': (self._process_order_by_clause, None),
+            'Token.Keyword.BY': (None, None),
+            'Token.Keyword.JOIN': (self._process_join, None),
+            'Token.Keyword.INNER': (self._process_join, None),
+            'Token.Keyword.LEFT': (self._process_join, None),
+            'Token.Keyword.RIGHT': (self._process_join, None),
+            'Token.Keyword.FULL': (self._process_join, None),
+            'Token.Keyword.OUTER': (self._process_join, None),
+            'Token.Keyword.CROSS': (self._process_join, None),
+            'Token.Keyword.ON': (self._process_on, None),
+            'Token.Keyword.AND': (self._process_and, None),
+            'Token.Keyword.OR': (self._process_or, None),
+            'Token.Keyword.IN': (self._process_in, None),
+            'Token.Keyword.LIKE': (self._process_like, None),
+            'Token.Keyword.BETWEEN': (self._process_between, None),
+            'Token.Keyword.IS': (self._process_is, None),
+            'Token.Keyword.AS': (self._process_as, (ST.SelectClause, ST.FromClause)),
+            'Token.Keyword.INTO': (self._process_into, None),
+            'Token.Keyword.NOT': (self._process_not, None),
+            'Token.Keyword.CASE': (self._process_case, None),
+            'Token.Keyword.WHEN': (self._process_when, None),
+            'Token.Keyword.THEN': (self._process_then, None),
+            'Token.Keyword.ELSE': (self._process_else, None),
+            'Token.Keyword.END': (self._process_end, None),
+            'Token.Keyword.UNION': (self._process_union, None),
+            'Token.Literal.Number': (self._process_literal_number, None),
+            'Token.Literal.String': (self._process_literal_string, None),
         }