+++ sqlsense/postgres/postgres_tokens.py
@@ -1,3 +1,10 @@
-from sqlsense.tokens import TokenGroup
+from pygments.token import Token
+
+from sqlsense.tokens import Identifier, TokenGroup
 
 LimitClause = TokenGroup.LimitClause
+
+WithClause = TokenGroup.WithClause
+WithIdentifier = TokenGroup.WithIdentifier
+WithQueryAliasIdentifier = Identifier.WithQueryAliasIdentifier
+WithQueryAliasName = Token.Name.WithQueryAliasName